fix: rename api folder
