+++ react-chat/src/components/Forms/RegistrationForm.jsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import useFormValidation from '../../hooks/useFormValidation';
 import validateProfileForm from '../../helpers/validateProfileForm';
 import ErrorMessage from '../../ErrorMessage/ErrorMessage.jsx';
-import registerUser from '../../API/USER/registerUser.js';
+import registerUser from '../../api/user/registerUser.js';
 import { useNavigate } from 'react-router-dom';
 import FormInput from '../FormElement/index.js';
 import Button from '../Button/Button.jsx';
