+++ react-chat/src/components/SendMessageForm/SendMessagesForm.module.scss
@@ -27,6 +27,9 @@
     border: 1px solid #00000033;
     outline: none;
     width: 100%;
+    &:focus-visible {
+      border: 2px solid $green-hover;
+    }
   }
   &__sendIcon {
     color: #ffffff;
