+++ react-chat/src/pages/MessageList/MessageList.jsx
@@ -5,10 +5,10 @@ import MessagesItem from '../../components/MessageItem/index.js';
 import { useNavigate, useParams } from 'react-router-dom';
 import useAutoScrollToBottom from '../../hooks/useAutoScrollToBottom.js';
 import { useLayoutEffect, useState } from 'react';
-import getMessages from '../../API/MESSAGES/getMessages.js';
-import getChatById from '../../API/CHAT/getChatById.js';
+import getMessages from '../../api/messages/getMessages.js';
+import getChatById from '../../api/chat/getChatById.js';
 import useCentrifuge from '../../hooks/useCentrifuge.js';
-import getCurrentUser from '../../API/USER/getCurrentUser.js';
+import getCurrentUser from '../../api/user/getCurrentUser.js';
 import createAvatar from '../../helpers/createAvatar.js';
 import Spinner from '../../components/Spinner/index.js';
 import { Player } from '@lottiefiles/react-lottie-player';
